refactor: 경로 조회 객체 별 역할 분리
